recall used in first stage classifier

+++ model.py
@@ -88,7 +88,8 @@ class Model(torch.nn.Module):
 
         # Compute loss if labels are provided
         if label is not None:
-          loss = torch.nn.functional.cross_entropy(input=logits, target=label)
-          output_dict['loss'] = loss
+            label = label.type(torch.LongTensor)
+            loss = torch.nn.functional.cross_entropy(input=logits, target=label)
+            output_dict['loss'] = loss
 
         return output_dict
